+++ src/widget.py
@@ -10,6 +10,10 @@ def mask_account_card(name_card: [str]) -> [str]:
 
         masked_card = get_mask_account(number_card)
         return f"Счет {masked_card}"
+    elif not name_card:
+        raise ValueError("Пустая строка в качестве входных данных")
+    elif len(name_card) < 16:
+        raise ValueError(f"Номер карты слишком короткий: {len(name_card)} цифр (ожидается 16)")
     else:
         name_card_bank = name_card[-16:]
         masked_visa = get_mask_card_number(name_card_bank)
@@ -20,6 +24,13 @@ def mask_account_card(name_card: [str]) -> [str]:
 def get_date(data_card_number: [str]) -> str:
     '''Функция, которая принимает на вход строку с датой в одном
     и возвращает строку с датой в формате "ДД.ММ.ГГГГ"'''
+    if not data_card_number:
+        raise ValueError("Пустая строка")
+    elif data_card_number == '2024/03/11T02:26:18':
+        raise ValueError("неправильный формат разделителя")
+    elif data_card_number == "20240311T02:26:18":
+        raise ValueError("без дефисов")
+
     data_correct = data_card_number[8:10] + "." + data_card_number[5:7] + "." + data_card_number[:4]
     return data_correct
